+++ src/main.js
@@ -1,5 +1,6 @@
 import { getSavedTalkIds, setSavedTalkIds, hasSaved } from "./LocalStorage";
 import "./loading.js";
+import { renderTopic, renderTalk } from "./Components.js";
 
 // inicio do main.js
 // criando um apelido para a função querySelectorAll
@@ -21,13 +22,22 @@ el.addEventListener('input', debounce(async function (el) {
   search(tab, query);
 }));
 
+el.addEventListener("keydown", function(event) {
+  if (event.key === "Enter") {
+      event.preventDefault();
+      search();
+  }
+});
+
 Array.from(tabs).forEach((radio) => {
   radio.addEventListener('change', function () {
     search(this.value, el.value);
   })
 })
-// fim do main.js
-// funcoes auxiliares
+
+window.handleToggleSave = handleToggleSave;
+
+
 
 function handleToggleSave(e) {
   const talkId = +e.dataset.id;
@@ -50,11 +60,13 @@ function handleToggleSave(e) {
   // Verifique a guia atual e atualize a exibição apenas na guia "saved"
   const activeTab = document.querySelector('input[name="tab"]:checked').value;
   if (activeTab === 'saved') {
-    search('saved', '');
+    search('saved', el.value);
   }
 }
+// fim do main.js
 // fim dos registros  eventos
 
+// funcoes auxiliares
 async function search(type = 'all', text = '') {
   // perguntar ao dados
   const data = await repositorioTalks(type);
@@ -156,78 +168,6 @@ async function repositorioTalks(type) {
   }
 }
 
-function renderTopic(card) {
-  return `<li class="topicos-cronograma-content">
-          <div>
-            <p id="horario-topico" class="horario-topico">${card.hour}</p>
-          </div>
-          <div class="nome-topico">
-            <p id="nome-topico">${card.title}</p>
-          </div>
-        </li>`;
-}
-
-
-function renderTalk(talk) {
-  const saved = hasSaved(talk);
-
-  // Mapeamento de substituição
-  const roomMappings = {
-    'principal': 'Auditório',
-    'invite': 'Convida',
-    'frontend': 'Front-End',
-    'communities': 'Comunidades',
-  };
-
-  // Verifique se o tipo de sala está no mapeamento e faça a substituição
-  const roomName = roomMappings[talk.room] || talk.room;
-
-  return `<li class="cards-palestrante-content">
-  <div class="info-post">
-      <div class="user">
-          <div class="user-img">
-              <img src="./assets/logos/frontenday2023.svg" alt="Logo Front-End Day 2023">
-          </div>
-          <div class="user-nome">
-              <p class="usuario">@frontendday2023</p>
-              <p class="horarioLive">${talk.hour}</p>
-          </div>
-      </div>
-  </div>
-  <div class="palestrante">
-  <div class="img-palestrante">
-  <img src="${talk.speaker.image}" alt="${talk.title}" class="imag-palestrante">
-  </div>
-  <div class="info-palestrante">
-  <div class="nome-cargo">
-  <h3 class="nomePalestrante">${talk.speaker.title}</h3>
-  <p class="cargoPalestrante">${talk.speaker.role} 
-  ${talk.speaker.company}</p>
-  </div>
-  
-  <div class="redes-profile">
-  <div class="linkedin-user">
-  <a href="${talk.speaker.social_link}" target="_blank"><i class="fa-brands fa-linkedin"></i>
-  <p class="linkedin">Linkedin</p>
-  </a>
-  </div>
-  </div>
-  
-  <div class="trilha">
-  <p class="trilhaPalestrante">Trilha: ${roomName}</p>
-  </div>
-  <div class="action">
-  <label><input type="checkbox" data-id="${talk.id}" ${saved ? 'checked' : ''} onclick="handleToggleSave(this)" /><span></span></label>
-</div>
-  </div>
-  </div>
-  <div class="assunto-palestra">
-  <p class="assuntoLive">${talk.title}</p>
-  </div>
-  
-</li>`;
-}
-
 // atrasando a execução de uma função: 
 // https://www.freecodecamp.org/portuguese/news/debounce-como-atrasar-a-execucao-de-uma-funcao-em-javascript-exemplo-com-js-es6/
 function debounce(fn, delay = 500) {
@@ -241,9 +181,9 @@ function debounce(fn, delay = 500) {
 
 window.handleToggleSave = handleToggleSave;
 
-el.addEventListener("keydown", function (event) {
-  if (event.key === "Enter") {
-    event.preventDefault();
-    search();
-  }
+el.addEventListener("keydown", function(event) {
+    if (event.key === "Enter") {
+        event.preventDefault();
+        search();
+    }
 });
